+++ src/main/java/com/zavier/classification/PaymentClassification.java
@@ -3,10 +3,14 @@ package com.zavier.classification;
 
 import com.zavier.Paycheck;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
+/**
+ * 计算薪水的方式(月薪、计时算薪水、底薪+销售提成等)
+ */
 public abstract class PaymentClassification {
-    public abstract double calculatePay(Paycheck pc);
+    public abstract BigDecimal calculatePay(Paycheck pc);
 
     public boolean isInPayPeriod(LocalDate theDate, Paycheck pc) {
         LocalDate payPeriodEndDate = pc.getPayPeriodEndDate();
